removed DeviceTimeWithRTC - don't want light updates to hang on I2C - RTOS loop should check for the time fault flag, and sync if needed
